Create newCandyPooler
Note: This script is attached to an empty gameObject named "CandyPooler." The script loads a set number of candy gameObjects into memory upon startup, and those gameObjects are immediately deactivated so that they are invisible to the player until they are called upon to appear, via the SpewCollectable script.

In the full game, candies are also set to be deactivated upon being eaten by the player. For the purpose of this demo, candies are deactivated after 30 seconds by the DestroyCandy script, and recycled back into the Object Pool for later use.